refactor: unify event processing architecture and message communication layer
Restructure messaging infrastructure to separate message queue operations from message API communication, and standardize event processing across all services
  using the Processor pattern.

Key changes:
- Rename messaging package to msgqueue for clarity and introduce MessageAPISender abstraction
- Replace EventListener pattern with ProcessorDispatcher pattern in message-api-server and message-dispatcher
- Consolidate relation operation handling to use integer flags instead of mixed types
- Add new /version endpoint and rename /relations to /friends in API gateway
- Introduce PayloadRouter for better message routing in message-dispatcher
- Add Push.proto for standardized push message definitions
- Fix typo in exception naming (NotFoundEvnetListenerException -> NotFoundListenerException)

+++ ripple-storage/src/main/java/com/fanaujie/ripple/storage/model/Relation.java
@@ -13,5 +13,5 @@ public class Relation {
     private String relationNickName;
     private String relationAvatar;
     private String relationRemarkName;
-    private int relationFlags; // bit0: is_friend, bit1: is_blocked, bit2: is_hidden
+    private byte relationFlags; // bit0: is_friend, bit1: is_blocked, bit2: is_hidden
 }
